Move refineUsingParent to Typer
